NU-2405 add config
