+++ designer/client/src/components/graph/node-modal/node/NodeContent/TestingContent.tsx
@@ -4,12 +4,9 @@ import React from "react";
 import { getUserSettings } from "../../../../../reducers/selectors/userSettings";
 import { useAppSelector } from "../../../../../store/storeHelpers";
 import type { NodeType } from "../../../../../types/node";
-import MockExpressionField from "../../editors/expression/MockExpressionField";
-import { NodeTable } from "../../NodeDetailsContent/NodeTable";
-import { getFindAvailableVariables } from "../../NodeDetailsContent/selectors";
-import { useGetNodeErrors, useIsEditMode, useSetProperty } from "../../useNodeTypeDetailsContentLogic";
 import type { NodeState } from "../useNodeState";
 import { InputData } from "./TestingContentElements/InputData";
+import { MockResponse } from "./TestingContentElements/mockResponse";
 
 interface Props {
     node: NodeType;
@@ -17,34 +14,31 @@ interface Props {
 }
 
 export const TestingContent = ({ node, onChange }: Props) => {
-    const settings = useAppSelector(getUserSettings);
-    const showMockFieldOnEnrichers = settings["node.showMockFieldOnEnrichers"];
+    const { getViewForNode } = useTestingContentRenderer();
+    const view = getViewForNode(node, onChange);
 
-    const showMockField = showMockFieldOnEnrichers && node.type === "Enricher" && node.service.id !== "decision-table";
-    const isEditMode = useIsEditMode({ onChange });
-    const setProperty = useSetProperty({ onChange, node });
-    const findAvailableVariables = useAppSelector(getFindAvailableVariables);
-    const [errors] = useGetNodeErrors(node);
-
-    if (node.type === "Source") {
-        return <InputData sourceId={node.id} />;
-    }
-
-    if (showMockField) {
-        return (
-            <NodeTable sx={settings["node.showInputsAndOutputs"] ? { margin: "0 16px" } : undefined}>
-                <MockExpressionField
-                    isEditMode={isEditMode}
-                    editedNode={node}
-                    showValidation
-                    showSwitch
-                    findAvailableVariables={findAvailableVariables}
-                    setNodeDataAt={setProperty}
-                    errors={errors}
-                />
-            </NodeTable>
-        );
-    }
-
-    return <Typography p={2}>No testing settings available for selected node</Typography>;
+    return view || <Typography p={2}>No testing settings available for selected node</Typography>;
 };
+
+export function useTestingContentRenderer() {
+    const settings = useAppSelector(getUserSettings);
+
+    const CONFIG = [
+        {
+            when: (node: NodeType) => node.type === "Source",
+            render: (node: NodeType) => <InputData sourceId={node.id} />,
+        },
+        {
+            when: (node: NodeType) =>
+                settings["node.showMockFieldOnEnrichers"] && node.type === "Enricher" && node.service.id !== "decision-table",
+            render: (node: NodeType, onChange?: NodeState["onChange"]) => <MockResponse node={node} onChange={onChange} />,
+        },
+    ];
+
+    const getViewForNode = (node: NodeType, onChange?: NodeState["onChange"]) => {
+        const matched = CONFIG.find((cfg) => cfg.when(node));
+        return matched?.render(node, onChange) ?? null;
+    };
+
+    return { getViewForNode };
+}
